Added role validation functionality

+++ node-with-sequelize/controller/regisUserController.js
@@ -27,9 +27,20 @@ const addRegisUser = async (payload, req, res) => {
 
 const getAllRegisUser = async (req, res) => {
   // console.log(`this is ${regisUser}`)
+  let queryIden = await RegisUser.findOne({ where: { id : req.body.id }})
 
-  let regisUser = await RegisUser.findAll({})
-  res.status(200).send(regisUser)
+  if (queryIden.role === "admin"){
+    let regisUser = await RegisUser.findAll({})
+    res.status(200).send(regisUser)
+  }
+  else{
+    return res.status(403).json({
+      success: false,
+      error: "Not authorized"
+    })
+  }
+
+  
 
 }
 
@@ -37,22 +48,43 @@ const getOneRegisUser = async (req, res) => {
 
   let id = req.params.id
   // console.log(id)
-  let regisUser = await RegisUser.findOne({ where: {id : id }})
-  res.status(200).send(regisUser)
-
+  let queryIden = await RegisUser.findOne({ where: { id : req.body.id }})
+  
+  
+  if (queryIden.role === 'admin'){
+    let regisUser = await RegisUser.findOne({ where: { id : id }})
+    res.status(200).send(regisUser)
+  }
+  else{
+    return res.status(403).json({
+      success: false,
+      error: "Not authorized"      
+    })
+  } 
 }
 
 const deleteRegisUser = async (req, res) => {
 
   let paramId = req.params.id
   // console.log(paramId)
-  await RegisUser.destroy({
-    where: {
-      id: paramId,
+  let queryIden = await RegisUser.findOne({ where: { id : req.body.id }})
+
+  if (queryIden.role === 'admin'){
+    await RegisUser.destroy({
+      where: {
+        id: paramId,
+      },
     },
-  },
-  );
-  res.status(200).send('record deleted')
+    );
+    res.status(200).send('record deleted')
+  }
+  else{
+    return res.status(403).json({
+      success: false,
+      error: "Not authorized"   
+    })
+  }
+
 
 }
 
@@ -60,12 +92,19 @@ const deleteRegisUser = async (req, res) => {
 const updateRegisUser = async (req, res) => {
 
   let id = req.params.id
+  let queryIden = await RegisUser.findOne({ where: { id : req.body.id }})
 
-  const regisUser = await RegisUser.update(req.body, { where: { id: id }})
+  if (queryIden.role === 'admin'){
+    const regisUser = await RegisUser.update(req.body.changed, { where: { id: id }})
 
-  res.status(200).send("record updated")
- 
-
+    res.status(200).send("record updated")
+  }
+  else{
+    return res.status(403).json({
+      success: false,
+      error: "Not authorized"   
+    })   
+  }
 }
